@@ -3,36 +3,31 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.8.0",
-    "@mui/material": "^5.8.1",
-    "@mui/styles": "^5.8.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
-    "cross-env": "^7.0.2",
-    "focus-trap-react": "^8.7.1",
-    "jquery": "^3.5.1",
-    "lodash": "^4.17.21",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.9.2",
+    "@mui/styles": "^5.9.2",
+    "axios": "^0.27.2",
+    "bootstrap": "^5.2.0",
+    "cross-env": "^7.0.3",
+    "focus-trap-react": "^9.0.2",
     "popper.js": "^1.14.3",
-    "qrcode.react": "^3.0.2",
-    "react": "^18.1.0",
+    "qrcode.react": "^3.1.0",
+    "react": "^18.2.0",
     "react-confetti": "^6.1.0",
-    "react-dom": "^18.1.0",
+    "react-dom": "^18.2.0",
     "react-input-mask": "^2.0.4",
     "react-linkify": "^1.0.0-alpha",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-router-hash-link": "^1.2.2",
-    "react-scripts": "^3.4.3",
-    "react-toastify": "^5.5.0",
-    "reactstrap": "^8.5.1",
-    "redux": "^4.0.5",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-router-hash-link": "^2.4.3",
+    "react-scripts": "^5.0.1",
+    "react-toastify": "^9.0.7",
+    "reactstrap": "^9.1.2",
+    "redux": "^4.2.0",
     "redux-persist": "^6.0.0",
-    "socket.io-client": "^4.0.2"
+    "socket.io-client": "^4.5.1"
   },
   "scripts": {
     "start": "cross-env BROWSER=false EXTEND_ESLINT=false SKIP_PREFLIGHT_CHECK=true react-scripts start",
